+++ apps/api/src/booking/booking.module.ts
@@ -1,15 +1,21 @@
 import { Module } from '@nestjs/common';
+
+import { BookingValidationModule } from '../booking-validation/booking-validation.module';
+import { RuleModule } from '../rule/rule.module';
+
+import { BookingReservationService } from './booking-reservation.service';
 import { BookingController } from './booking.controller';
 import { BookingService } from './booking.service';
-import { RuleModule } from '../rule/rule.module';
-import { BookingReservationService } from './booking-reservation.service';
 
 @Module({
-  imports: [RuleModule],
+  imports: [
+    RuleModule,
+    BookingValidationModule,
+  ],
   controllers: [BookingController],
-providers: [
-  BookingService,
-  BookingReservationService,
-],
+  providers: [
+    BookingService,
+    BookingReservationService,
+  ],
 })
 export class BookingModule {}